+++ lib/core/models/editor_callbacks/main_editor/main_editor_callbacks.dart
@@ -2,6 +2,7 @@
 import 'package:flutter/widgets.dart';
 
 import '/features/main_editor/main_editor.dart';
+import '/features/main_editor/services/state_manager.dart';
 import '/shared/services/import_export/import_state_history.dart';
 import '../../../enums/sub_editors_name.dart';
 import '../../layers/layer.dart';
@@ -40,6 +41,8 @@ class MainEditorCallbacks extends StandaloneEditorCallbacks {
     this.onImportHistoryStart,
     this.onImportHistoryEnd,
     this.onHoverRemoveAreaChange,
+    this.onStateHistoryChange,
+    this.onImageDecoded,
     super.onInit,
     super.onAfterViewInit,
     super.onUpdateUI,
@@ -99,6 +102,14 @@ class MainEditorCallbacks extends StandaloneEditorCallbacks {
   /// This can be `null` if no action is required at the start of the close
   /// process.
   final Function(SubEditor editor)? onStartCloseSubEditor;
+
+  /// Callback that is triggered whenever the state history of the editor
+  /// changes.
+  final Function(StateManager stateHistory, ProImageEditorState editor)?
+      onStateHistoryChange;
+
+  /// Callback that is triggered after the image has been successfully decoded.
+  final Function()? onImageDecoded;
 
   /// A callback function that is triggered when the user `tap` on the body.
   final Function()? onTap;
@@ -406,11 +417,14 @@ class MainEditorCallbacks extends StandaloneEditorCallbacks {
     Function()? onDone,
     Function()? onRedo,
     Function()? onUndo,
+    Function()? onImageDecoded,
     Function(ProImageEditorState state, ImportStateHistory import)?
         onImportHistoryStart,
     Function(ProImageEditorState state, ImportStateHistory import)?
         onImportHistoryEnd,
     Function(bool isPointerInside)? onHoverRemoveAreaChange,
+    Function(StateManager stateHistory, ProImageEditorState editor)?
+        onStateHistoryChange,
   }) {
     return MainEditorCallbacks(
       onLayerTapDown: onLayerTapDown ?? this.onLayerTapDown,
@@ -447,10 +461,12 @@ class MainEditorCallbacks extends StandaloneEditorCallbacks {
       onDone: onDone ?? this.onDone,
       onRedo: onRedo ?? this.onRedo,
       onUndo: onUndo ?? this.onUndo,
+      onImageDecoded: onImageDecoded ?? this.onImageDecoded,
       onImportHistoryStart: onImportHistoryStart ?? this.onImportHistoryStart,
       onImportHistoryEnd: onImportHistoryEnd ?? this.onImportHistoryEnd,
       onHoverRemoveAreaChange:
           onHoverRemoveAreaChange ?? this.onHoverRemoveAreaChange,
+      onStateHistoryChange: onStateHistoryChange ?? this.onStateHistoryChange,
     );
   }
 }